requirements.txt
djangorestframework
django-filter
python-dateutil
django-ckeditor
django-summernote
django-gcm
django-multiupload